begin implementation of plugin data
Started on the process of importing plugin data from JF Reporting Plugin
Data is stored in a separate table and will be used in a new view to allow for this data to work in conjunction with the data captured by JFStat. These new views will be finalized and released in the next push.

Reworked watchdog function to use a unique uuid instead of the session id as it is not unique and seems to be causing issues regardless of no pk constraints
Also decided to use moments to get accurate timezone data

Changed item info to sync only items without item info. This is temporary to reduce the load of subsequent syncs until we can optimize this function

Minor changes to front end ui linkages and added open on jellyfin link on item details page