Fix for sync issues
Change to Included mixed libraries broke sync as Jellyfin does not populate a CollectionType for this library type breaking the no-null constraint. A default value of 'mixed' will be passed instead of a null.

Modified js_library_stats_overview to account for empty libraries

Modified Libraries view to cater for the above changes + added error boundaries