Playback Reporting Plugin
Added support for the Playback Reporting Plugin.

This will only import data from this table that is older then the first record captured by Jellystat to prevent duplication.

This means that it will not import any data going forward. its only meant to import old data to have parity with your old watch logs.

Due to the limited set of information logged by the plugin, you may notice some information that's missing when compared to data logged by Jellystat.

Other changes:
ServerId column added as the foundation for future updates that may or may not include multiple server support. This addition is the most crucial and is why it was added now and not held for future change sets.