added archiving feature to libraries
added archiving flag to libraries. If a library is deleted or excluded the entire library and its contents are archived when the sync process is run

Added archive toggle in libraries to show/hide archived libraries

Still need to add functionality to bulk purge archived data for a library or for selected items in the library.

cleaned up duplicate statement in migration 46

uncommented backwards navigation code when purge completes