Fixed Partial Sync issues
Fixed some logic that was causing show items to sync during partial sync but then get removed when the clean-up script was run.

This was due to the Show filter used to reduce the number of shows scanned but a scenario was found where:

Recently added returns an episode and not the corresponding series item.  The filter would then prevent the episode from syncing as the filter excluded it based on the missing Show ID.  Then once the item info sync ran, the item info would insert without the corresponding item id.

The cleanup looks for this scenario where theres info items without corresponding items and cleans them out.